disable serial console, gokr-packer will re-enable
The uart0, uart1, serial0 and serial1 aliases are required to make the
Raspberry Pi firmware files not rewrite the kernel command line.

With that out of the way, we can move the logic for enabling or
disabling the serial port into gokr-packer, which modifies config.txt
and cmdline.txt accordingly.

related to https://github.com/gokrazy/gokrazy/issues/11